requirements.txt
html-table-parser-python3
mysql-connector-python
python-dotenv
selenium
webdriver-manager
pandas
folium
geopy
pycountry

